+++ src/main/webapp/components/confirm-delete/confirm-delete.js
