+++ src/main/webapp/components/confirm-delete/confirm-delete.stories.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { storiesOf } from '../@storybook/react'
+import { storiesOf } from '../../@storybook/react'
 
 import { text } from '@connexta/ace/@storybook/addon-knobs'
