+++ src/main/webapp/components/confirm-delete/index.js
