+++ src/main/webapp/components/index-cards/index-cards.js
@@ -15,7 +15,7 @@ import CardActionArea from '@material-ui/core/CardActionArea'
 
 import moment from 'moment'
 
-import ConfirmDelete from '../../confirm-delete'
+import ConfirmDelete from '../confirm-delete'
 import SharingModal from '../../sharing/sharing-modal'
 
 const onClick = action => e => {
